# mumble_link_server/openapi/openapi.yaml
openapi: 3.0.3
info:
  description: OpenAPI spec for mumble link server and client
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  title: Mumble Link API
  version: 1.0.0
servers:
- url: http://localhost:8080
security:
- Bearer: []
paths:
  /server/create:
    post:
      operationId: server_create_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_server_create_post_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
          description: Server
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Create a new server
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
  /server/{server_id}:
    delete:
      description: Delete a server
      operationId: server_server_id_delete
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          description: Empty
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
    get:
      description: Get server info
      operationId: server_server_id_get
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
          description: Server
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
    put:
      description: Modify server parameters
      operationId: server_server_id_put
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_server__server_id__put_request'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
          description: Server
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
  /server/{server_id}/channel/create:
    post:
      description: Create a new channel
      operationId: server_server_id_channel_create_post
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_server__server_id__channel_create_post_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
          description: Channel
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
  /server/{server_id}/channel/{channel_id}:
    delete:
      description: Delete a channel
      operationId: server_server_id_channel_channel_id_delete
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      - in: path
        name: channel_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: Empty
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
    get:
      description: Get channel info
      operationId: server_server_id_channel_channel_id_get
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      - in: path
        name: channel_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
          description: Channel
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
    put:
      description: Modify channel parameters
      operationId: server_server_id_channel_channel_id_put
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      - in: path
        name: channel_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_server__server_id__channel__channel_id__put_request'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
          description: Channel
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
  /server/{server_id}/channels:
    get:
      description: Get list of channels
      operationId: server_server_id_channels_get
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Channel'
                type: array
          description: Channels
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
  /server/{server_id}/group/create:
    post:
      description: Create a new group
      operationId: server_server_id_group_create_post
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_server__server_id__group_create_post_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
          description: Group
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
  /server/{server_id}/group/{group_id}:
    delete:
      description: Delete a group
      operationId: server_server_id_group_group_id_delete
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      - in: path
        name: group_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          description: Empty
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
    get:
      description: Get group info
      operationId: server_server_id_group_group_id_get
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      - in: path
        name: group_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
          description: Group
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
    put:
      description: Modify group parameters
      operationId: server_server_id_group_group_id_put
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      - in: path
        name: group_id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_server__server_id__group__group_id__put_request'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
          description: Group
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
  /server/{server_id}/groups:
    get:
      description: Get list of groups
      operationId: server_server_id_groups_get
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Group'
                type: array
          description: Groups
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
  /server/{server_id}/user/{user_id}:
    get:
      description: Get user info
      operationId: server_server_id_user_user_id_get
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      - in: path
        name: user_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: User
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
  /server/{server_id}/user/{user_id}/ban:
    post:
      description: Ban a user
      operationId: server_server_id_user_user_id_ban_post
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      - in: path
        name: user_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_server__server_id__user__user_id__ban_post_request'
        required: true
      responses:
        "200":
          description: Empty
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
  /server/{server_id}/user/{user_id}/bans:
    get:
      description: Get list of banned users
      operationId: server_server_id_user_user_id_bans_get
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ban'
                type: array
          description: Bans
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
  /server/{server_id}/user/{user_id}/kick:
    post:
      description: Kick a user
      operationId: server_server_id_user_user_id_kick_post
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      - in: path
        name: user_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/_server__server_id__user__user_id__kick_post_request'
        required: true
      responses:
        "200":
          description: Empty
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
...
  /server/{server_id}/user/{user_id}/mute:
    post:
      description: Mute a user
      operationId: server_server_id_user_user_id_mute_post
      parameters:
      - in: path
        name: server_id
        required: true
        schema:
          format: uuid
          type: string
      - in: path
        name: user_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: Empty
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      x-openapi-router-controller: mumble_link_server.controllers.default_controller
...
components:
...
  schemas:
    Server:
      example:
        server_name: server_name
        user_count: 6
        superuser_password: superuser_password
        server_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        server_password: server_password
        max_users: 0
      properties:
        server_id:
          format: uuid
          title: server_id
          type: string
        server_name:
          title: server_name
          type: string
        max_users:
          title: max_users
          type: integer
        user_count:
          title: user_count
          type: integer
        server_password:
          title: server_password
          type: string
        superuser_password:
          title: superuser_password
          type: string
      required:
      - max_users
      - server_id
      - server_name
      - user_count
      title: Server
      type: object
...
    _server_create_post_request:
      properties:
        name:
          title: name
          type: string
        max_users:
          title: max_users
          type: integer
        server_password:
          title: server_password
          type: string
        superuser_password:
          title: superuser_password
          type: string
      required:
      - max_users
      - name
      title: _server_create_post_request
      type: object
...
  securitySchemes:
    Bearer:
      scheme: bearer
      type: http
      x-bearerInfoFunc: mumble_link_server.controllers.security_controller_.info_from_Bearer